Code review: 300680043: Split multi_processing into separate files.
