+++ config/dpkg/changelog
@@ -2,4 +2,4 @@ plaso (1.4.1-1) unstable; urgency=low
 
   * Auto-generated
 
- -- Log2Timeline <log2timeline-dev@googlegroups.com>  Sat, 09 Jul 2016 17:00:08 +0200
+ -- Log2Timeline <log2timeline-dev@googlegroups.com>  Sun, 10 Jul 2016 09:46:54 +0200
