+++ docs/plaso.multi_processing.rst
@@ -4,10 +4,26 @@ plaso.multi_processing package
 Submodules
 ----------
 
-plaso.multi_processing.multi_process module
--------------------------------------------
+plaso.multi_processing.base_process module
+------------------------------------------
 
-.. automodule:: plaso.multi_processing.multi_process
+.. automodule:: plaso.multi_processing.base_process
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+plaso.multi_processing.engine module
+------------------------------------
+
+.. automodule:: plaso.multi_processing.engine
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+plaso.multi_processing.multi_process_queue module
+-------------------------------------------------
+
+.. automodule:: plaso.multi_processing.multi_process_queue
     :members:
     :undoc-members:
     :show-inheritance:
@@ -32,6 +48,14 @@ plaso.multi_processing.task_manager module
 ------------------------------------------
 
 .. automodule:: plaso.multi_processing.task_manager
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+plaso.multi_processing.worker_process module
+--------------------------------------------
+
+.. automodule:: plaso.multi_processing.worker_process
     :members:
     :undoc-members:
     :show-inheritance:
