+++ plaso/multi_processing/engine.py
@@ -1,19 +1,16 @@
 # -*- coding: utf-8 -*-
 """The multi-process processing engine."""
 
-import abc
 import ctypes
 import logging
 import multiprocessing
 import os
 import Queue
-import random
 import signal
 import sys
 import threading
 import time
 
-from dfvfs.lib import errors as dfvfs_errors
 from dfvfs.resolver import context
 
 from plaso.containers import event_sources
@@ -21,207 +18,13 @@ from plaso.engine import engine
 from plaso.engine import extractors
 from plaso.engine import plaso_queue
 from plaso.engine import profiler
-from plaso.engine import worker
 from plaso.engine import zeromq_queue
 from plaso.lib import definitions
-from plaso.lib import errors
+from plaso.multi_processing import multi_process_queue
 from plaso.multi_processing import process_info
 from plaso.multi_processing import task_manager
+from plaso.multi_processing import worker_process
 from plaso.multi_processing import xmlrpc
-from plaso.parsers import mediator as parsers_mediator
-
-
-class MultiProcessBaseProcess(multiprocessing.Process):
-  """Class that defines the multi-processing process interface.
-
-  Attributes:
-    rpc_port (int): port number of the process status RPC server.
-  """
-
-  _NUMBER_OF_RPC_SERVER_START_ATTEMPTS = 14
-  _PROCESS_JOIN_TIMEOUT = 5.0
-
-  def __init__(self, enable_sigsegv_handler=False, **kwargs):
-    """Initializes a process object.
-
-    Args:
-      enable_sigsegv_handler (bool): True if the SIGSEGV handler should
-                                     be enabled.
-      kwargs: keyword arguments to pass to multiprocessing.Process.
-    """
-    super(MultiProcessBaseProcess, self).__init__(**kwargs)
-    self._enable_sigsegv_handler = enable_sigsegv_handler
-    self._original_sigsegv_handler = None
-    # TODO: check if this can be replaced by self.pid or does this only apply
-    # to the parent process?
-    self._pid = None
-    self._rpc_server = None
-    self._status_is_running = False
-
-    # We need to share the RPC port number with the engine process.
-    self.rpc_port = multiprocessing.Value(u'I', 0)
-
-  @property
-  def name(self):
-    """The process name."""
-    return self._name
-
-  @abc.abstractmethod
-  def _GetStatus(self):
-    """Returns a status dictionary."""
-
-  @abc.abstractmethod
-  def _Main(self):
-    """The process main loop.
-
-    This method is called when the process is ready to start. A sub class
-    should override this method to do the necessary actions in the main loop.
-    """
-
-  def _OnCriticalError(self):
-    """The process on critical error handler.
-
-    This method is called when the process encountered a critical error e.g.
-    a segfault. A sub class should override this method to do the necessary
-    actions before the original critical error signal handler it called.
-
-    Be aware that the state of the process should not be trusted a significant
-    part of memory could have been overwritten before a segfault. This callback
-    is primarily intended to salvage what we need to troubleshoot the error.
-    """
-    return
-
-  def _SigSegvHandler(self, unused_signal_number, unused_stack_frame):
-    """Signal handler for the SIGSEGV signal.
-
-    Args:
-      signal_number: Numeric representation of the signal.
-      stack_frame: The current stack frame (instance of frame object) or None.
-    """
-    self._OnCriticalError()
-
-    # Note that the original SIGSEGV handler can be 0.
-    if self._original_sigsegv_handler is not None:
-      # Let the original SIGSEGV handler take over.
-      signal.signal(signal.SIGSEGV, self._original_sigsegv_handler)
-      os.kill(self._pid, signal.SIGSEGV)
-
-  def _SigTermHandler(self, unused_signal_number, unused_stack_frame):
-    """Signal handler for the SIGTERM signal.
-
-    Args:
-      signal_number: Numeric representation of the signal.
-      stack_frame: The current stack frame (instance of frame object) or None.
-    """
-    self.SignalAbort()
-
-  def _StartProcessStatusRPCServer(self):
-    """Starts the process status RPC server."""
-    if self._rpc_server:
-      return
-
-    self._rpc_server = xmlrpc.XMLProcessStatusRPCServer(self._GetStatus)
-
-    hostname = u'localhost'
-
-    # Try the PID as port number first otherwise pick something random
-    # between 1024 and 60000.
-    if self._pid < 1024 or self._pid > 60000:
-      port = random.randint(1024, 60000)
-    else:
-      port = self._pid
-
-    if not self._rpc_server.Start(hostname, port):
-      port = 0
-      for _ in range(self._NUMBER_OF_RPC_SERVER_START_ATTEMPTS):
-        port = random.randint(1024, 60000)
-        if self._rpc_server.Start(hostname, port):
-          break
-
-        port = 0
-
-    if not port:
-      logging.error((
-          u'Unable to start a process status RPC server for {0!s} '
-          u'(PID: {1:d})').format(self._name, self._pid))
-      self._rpc_server = None
-      return
-
-    self.rpc_port.value = port
-
-    logging.debug(
-        u'Process: {0!s} process status RPC server started'.format(self._name))
-
-  def _StopProcessStatusRPCServer(self):
-    """Stops the process status RPC server."""
-    if not self._rpc_server:
-      return
-
-    # Make sure the engine gets one more status update so it knows
-    # the worker has completed.
-    self._WaitForStatusNotRunning()
-
-    self._rpc_server.Stop()
-    self._rpc_server = None
-    self.rpc_port.value = 0
-
-    logging.debug(
-        u'Process: {0!s} process status RPC server stopped'.format(self._name))
-
-  def _WaitForStatusNotRunning(self):
-    """Waits for the status is running to change to false."""
-    # We wait slightly longer than the status check sleep time.
-    time.sleep(2.0)
-    time_slept = 2.0
-    while self._status_is_running:
-      time.sleep(0.5)
-      time_slept += 0.5
-      if time_slept >= self._PROCESS_JOIN_TIMEOUT:
-        break
-
-  # This method is part of the multiprocessing.Process interface hence
-  # its name does not follow the style guide.
-  def run(self):
-    """Runs the process."""
-    # Prevent the KeyboardInterrupt being raised inside the process.
-    # This will prevent a process to generate a traceback when interrupted.
-    signal.signal(signal.SIGINT, signal.SIG_IGN)
-
-    # A SIGTERM signal handler is necessary to make sure IPC is cleaned up
-    # correctly on terminate.
-    signal.signal(signal.SIGTERM, self._SigTermHandler)
-
-    # A SIGSEGV signal handler is necessary to try to indicate where
-    # worker failed.
-    # WARNING the SIGSEGV handler will deadlock the process on a real segfault.
-    if self._enable_sigsegv_handler:
-      self._original_sigsegv_handler = signal.signal(
-          signal.SIGSEGV, self._SigSegvHandler)
-
-    self._pid = os.getpid()
-
-    # We need to set the is running status explicitly to True in case
-    # the process completes before the engine is able to determine
-    # the status of the process, e.g. in the unit tests.
-    self._status_is_running = True
-
-    logging.debug(
-        u'Process: {0!s} (PID: {1:d}) started'.format(self._name, self._pid))
-
-    self._StartProcessStatusRPCServer()
-
-    self._Main()
-
-    self._StopProcessStatusRPCServer()
-
-    logging.debug(
-        u'Process: {0!s} (PID: {1:d}) stopped'.format(self._name, self._pid))
-
-    self._status_is_running = False
-
-  @abc.abstractmethod
-  def SignalAbort(self):
-    """Signals the process to abort."""
 
 
 class MultiProcessEngine(engine.BaseEngine):
@@ -419,8 +222,9 @@ class MultiProcessEngine(engine.BaseEngine):
 
       logging.info(u'Starting replacement worker process for {0:s}'.format(
           process.name))
-      worker_process = self._StartExtractionWorkerProcess(self._storage_writer)
-      self._StartMonitoringProcess(worker_process.pid)
+      replacement_process = self._StartExtractionWorkerProcess(
+          self._storage_writer)
+      self._StartMonitoringProcess(replacement_process.pid)
 
     elif status_indicator == definitions.PROCESSING_STATUS_COMPLETED:
       number_of_events = process_status.get(u'number_of_events', 0)
@@ -722,7 +526,7 @@ class MultiProcessEngine(engine.BaseEngine):
     else:
       task_queue = self._task_queue
 
-    worker_process = MultiProcessWorkerProcess(
+    process = worker_process.WorkerProcess(
         task_queue, storage_writer, self.knowledge_base,
         self._session_identifier, self._last_worker_number,
         enable_debug_output=self._enable_debug_output,
@@ -739,12 +543,12 @@ class MultiProcessEngine(engine.BaseEngine):
         temporary_directory=self._temporary_directory,
         text_prepend=self._text_prepend)
 
-    worker_process.start()
+    process.start()
     self._last_worker_number += 1
 
-    self._RegisterProcess(worker_process)
+    self._RegisterProcess(process)
 
-    return worker_process
+    return process
 
   def _StartMonitoringProcess(self, pid):
     """Starts monitoring a process.
@@ -860,7 +664,7 @@ class MultiProcessEngine(engine.BaseEngine):
 
     # For Multiprocessing queues, set abort to True to stop queue.join_thread()
     # from blocking.
-    if isinstance(self._task_queue, MultiProcessingQueue):
+    if isinstance(self._task_queue, multi_process_queue.MultiProcessingQueue):
       self._task_queue.Close(abort=True)
 
     if abort:
@@ -1089,7 +893,7 @@ class MultiProcessEngine(engine.BaseEngine):
 
     # Set up the task queue.
     if not self._use_zeromq:
-      self._task_queue = MultiProcessingQueue(
+      self._task_queue = multi_process_queue.MultiProcessingQueue(
           maximum_number_of_queued_items=self._maximum_number_of_tasks)
 
     else:
@@ -1177,446 +981,3 @@ class MultiProcessEngine(engine.BaseEngine):
     self._text_prepend = None
 
     return self._processing_status
-
-
-class MultiProcessWorkerProcess(MultiProcessBaseProcess):
-  """Class that defines a multi-processing worker process."""
-
-  def __init__(
-      self, task_queue, storage_writer, knowledge_base, session_identifier,
-      worker_number, enable_debug_output=False, enable_profiling=False,
-      filter_object=None, hasher_names_string=None, mount_path=None,
-      parser_filter_expression=None, process_archive_files=False,
-      profiling_directory=None, profiling_sample_rate=1000,
-      profiling_type=u'all', temporary_directory=None, text_prepend=None,
-      **kwargs):
-    """Initializes a process object.
-
-    Non-specified keyword arguments (kwargs) are directly passed to
-    multiprocessing.Process.
-
-    Args:
-      task_queue (MultiProcessingQueue): task queue.
-      storage_writer (StorageWriter): storage writer for a session storage.
-      knowledge_base (KnowledgeBase): knowledge base which contains
-          information from the source data needed for parsing.
-      session_identifier (str): identifier of the session.
-      worker_number: a number that identifies the worker.
-      enable_debug_output (Optional[bool]): True if debug output should be
-          enabled.
-      enable_profiling (Optional[bool]): True if profiling should be enabled.
-      filter_object (Optional[objectfilter.Filter]): filter object.
-      hasher_names_string (Optional[str]): comma separated string of names
-          of hashers to use during processing.
-      mount_path (Optional[str]): mount path.
-      parser_filter_expression (Optional[str]): parser filter expression,
-          where None represents all parsers and plugins.
-      process_archive_files (Optional[bool]): True if archive files should be
-          scanned for file entries.
-      profiling_directory (Optional[str]): path to the directory where
-          the profiling sample files should be stored.
-      profiling_sample_rate (Optional[int]): the profiling sample rate.
-          Contains the number of event sources processed.
-      profiling_type (Optional[str]): type of profiling.
-          Supported types are:
-
-          * 'memory' to profile memory usage;
-          * 'parsers' to profile CPU time consumed by individual parsers;
-          * 'processing' to profile CPU time consumed by different parts of
-            the processing;
-          * 'serializers' to profile CPU time consumed by individual
-            serializers.
-      temporary_directory (Optional[str]): path of the directory for temporary
-          files.
-      text_prepend (Optional[str]): text to prepend to every event.
-      kwargs: keyword arguments to pass to multiprocessing.Process.
-    """
-    super(MultiProcessWorkerProcess, self).__init__(**kwargs)
-    self._abort = False
-    self._buffer_size = 0
-    self._enable_debug_output = enable_debug_output
-    self._enable_profiling = enable_profiling
-    self._extraction_worker = None
-    self._filter_object = filter_object
-    self._hasher_names_string = hasher_names_string
-    self._knowledge_base = knowledge_base
-    self._memory_profiler = None
-    self._mount_path = mount_path
-    self._number_of_consumed_events = 0
-    self._number_of_consumed_sources = 0
-    self._parser_filter_expression = parser_filter_expression
-    self._parser_mediator = None
-    self._parsers_profiler = None
-    self._processing_profiler = None
-    self._serializers_profiler = None
-    self._process_archive_files = process_archive_files
-    self._profiling_directory = profiling_directory
-    self._profiling_sample_rate = profiling_sample_rate
-    self._profiling_type = profiling_type
-    self._session_identifier = session_identifier
-    self._status = definitions.PROCESSING_STATUS_INITIALIZED
-    self._storage_writer = storage_writer
-    self._task_identifier = u''
-    self._task_queue = task_queue
-    self._temporary_directory = temporary_directory
-    self._text_prepend = text_prepend
-    self._worker_number = worker_number
-
-  def _GetStatus(self):
-    """Returns a status dictionary."""
-    if self._parser_mediator:
-      number_of_produced_errors = (
-          self._parser_mediator.number_of_produced_errors)
-      number_of_produced_events = (
-          self._parser_mediator.number_of_produced_events)
-      number_of_produced_sources = (
-          self._parser_mediator.number_of_produced_event_sources)
-    else:
-      number_of_produced_errors = 0
-      number_of_produced_events = 0
-      number_of_produced_sources = 0
-
-    if self._extraction_worker:
-      current_display_name = self._extraction_worker.current_display_name
-      processing_status = self._extraction_worker.processing_status
-    else:
-      current_display_name = u''
-      processing_status = self._status
-
-    status = {
-        u'display_name': current_display_name,
-        u'identifier': self._name,
-        u'number_of_consumed_errors': 0,
-        u'number_of_consumed_events': self._number_of_consumed_events,
-        u'number_of_consumed_sources': self._number_of_consumed_sources,
-        u'number_of_produced_errors': number_of_produced_errors,
-        u'number_of_produced_events': number_of_produced_events,
-        u'number_of_produced_sources': number_of_produced_sources,
-        u'processing_status': processing_status,
-        u'task_identifier': self._task_identifier}
-
-    self._status_is_running = status.get(u'is_running', False)
-    return status
-
-  def _Main(self):
-    """The main loop."""
-    self._parser_mediator = parsers_mediator.ParserMediator(
-        None, self._knowledge_base,
-        temporary_directory=self._temporary_directory)
-
-    if self._filter_object:
-      self._parser_mediator.SetFilterObject(self._filter_object)
-
-    if self._mount_path:
-      self._parser_mediator.SetMountPath(self._mount_path)
-
-    if self._text_prepend:
-      self._parser_mediator.SetTextPrepend(self._text_prepend)
-
-    # We need a resolver context per process to prevent multi processing
-    # issues with file objects stored in images.
-    resolver_context = context.Context()
-
-    # We need to initialize the parser and hasher objects after the process
-    # has forked otherwise on Windows the "fork" will fail with
-    # a PickleError for Python modules that cannot be pickled.
-    self._extraction_worker = worker.EventExtractionWorker(
-        resolver_context,
-        parser_filter_expression=self._parser_filter_expression,
-        process_archive_files=self._process_archive_files)
-
-    if self._hasher_names_string:
-      self._extraction_worker.SetHashers(self._hasher_names_string)
-
-    self._StartProfiling()
-
-    logging.debug(u'Worker: {0!s} (PID: {1:d}) started'.format(
-        self._name, self._pid))
-
-    self._status = definitions.PROCESSING_STATUS_RUNNING
-
-    try:
-      logging.debug(
-          u'{0!s} (PID: {1:d}) started monitoring task queue.'.format(
-              self._name, self._pid))
-
-      while not self._abort:
-        try:
-          task = self._task_queue.PopItem()
-        except (errors.QueueClose, errors.QueueEmpty) as exception:
-          logging.debug(u'ConsumeItems exiting with exception {0:s}.'.format(
-              type(exception)))
-          break
-
-        if isinstance(task, plaso_queue.QueueAbort):
-          logging.debug(u'ConsumeItems exiting, dequeued QueueAbort object.')
-          break
-
-        self._ProcessTask(task)
-
-      logging.debug(
-          u'{0!s} (PID: {1:d}) stopped monitoring task queue.'.format(
-              self._name, self._pid))
-
-    except Exception as exception:  # pylint: disable=broad-except
-      logging.warning(
-          u'Unhandled exception in worker: {0!s} (PID: {1:d}).'.format(
-              self._name, self._pid))
-      logging.exception(exception)
-
-      self._abort = True
-
-    self._StopProfiling()
-    self._extraction_worker = None
-    self._parser_mediator = None
-
-    if self._abort:
-      self._status = definitions.PROCESSING_STATUS_ABORTED
-    else:
-      self._status = definitions.PROCESSING_STATUS_COMPLETED
-
-    logging.debug(u'Extraction worker: {0!s} (PID: {1:d}) stopped'.format(
-        self._name, self._pid))
-
-    if isinstance(self._task_queue, MultiProcessingQueue):
-      self._task_queue.Close(abort=True)
-    else:
-      self._task_queue.Close()
-
-  def _ProcessPathSpec(self, extraction_worker, parser_mediator, path_spec):
-    """Processes a path specification.
-
-    Args:
-      extraction_worker (worker.ExtractionWorker): extraction worker.
-      parser_mediator (ParserMediator): parser mediator.
-      path_spec (dfvfs.PathSpec): path specification.
-    """
-    try:
-      extraction_worker.ProcessPathSpec(parser_mediator, path_spec)
-
-    except IOError as exception:
-      logging.warning((
-          u'Unable to process path specification: {0:s} with error: '
-          u'{1:s}').format(extraction_worker.current_display_name, exception))
-
-    except dfvfs_errors.CacheFullError:
-      # TODO: signal engine of failure.
-      self._abort = True
-      logging.error((
-          u'ABORT: detected cache full error while processing '
-          u'path spec: {0:s}').format(
-              extraction_worker.current_display_name))
-
-    # All exceptions need to be caught here to prevent the worker
-    # from being killed by an uncaught exception.
-    except Exception as exception:  # pylint: disable=broad-except
-      logging.warning(
-          u'Unhandled exception while processing path spec: {0:s}.'.format(
-              extraction_worker.current_display_name))
-      logging.exception(exception)
-
-  def _ProcessTask(self, task):
-    """Processes a task.
-
-    Args:
-      task (Task): task.
-    """
-    self._task_identifier = task.identifier
-
-    storage_writer = self._storage_writer.CreateTaskStorage(task)
-
-    if self._serializers_profiler:
-      storage_writer.SetSerializersProfiler(self._serializers_profiler)
-
-    storage_writer.Open()
-
-    try:
-      self._parser_mediator.SetStorageWriter(storage_writer)
-
-      storage_writer.WriteTaskStart()
-
-      # TODO: add support for more task types.
-      self._ProcessPathSpec(
-          self._extraction_worker, self._parser_mediator, task.path_spec)
-      self._number_of_consumed_sources += 1
-
-      if self._memory_profiler:
-        self._memory_profiler.Sample()
-
-      # TODO: on abort use WriteTaskAborted instead of completion?
-      storage_writer.WriteTaskCompletion()
-
-    finally:
-      self._parser_mediator.SetStorageWriter(None)
-
-      storage_writer.Close()
-
-      if self._serializers_profiler:
-        storage_writer.SetSerializersProfiler(None)
-
-    self._storage_writer.PrepareMergeTaskStorage(task.identifier)
-
-    self._task_identifier = u''
-
-  def _StartProfiling(self):
-    """Starts profiling."""
-    if not self._enable_profiling:
-      return
-
-    if self._profiling_type in (u'all', u'memory'):
-      identifier = u'{0:s}-memory'.format(self._name)
-      self._memory_profiler = profiler.GuppyMemoryProfiler(
-          identifier, path=self._profiling_directory,
-          profiling_sample_rate=self._profiling_sample_rate)
-      self._memory_profiler.Start()
-
-    if self._profiling_type in (u'all', u'parsers'):
-      identifier = u'{0:s}-parsers'.format(self._name)
-      self._parsers_profiler = profiler.ParsersProfiler(
-          identifier, path=self._profiling_directory)
-      self._extraction_worker.SetParsersProfiler(self._parsers_profiler)
-
-    if self._profiling_type in (u'all', u'processsing'):
-      identifier = u'{0:s}-processsing'.format(self._name)
-      self._processing_profiler = profiler.ProcessingProfiler(
-          identifier, path=self._profiling_directory)
-      self._extraction_worker.SetProcessingProfiler(self._processing_profiler)
-
-    if self._profiling_type in (u'all', u'serializers'):
-      identifier = u'{0:s}-serializers'.format(self._name)
-      self._serializers_profiler = profiler.SerializersProfiler(
-          identifier, path=self._profiling_directory)
-
-  def _StopProfiling(self):
-    """Stops profiling."""
-    if not self._enable_profiling:
-      return
-
-    if self._profiling_type in (u'all', u'memory'):
-      self._memory_profiler.Sample()
-      self._memory_profiler = None
-
-    if self._profiling_type in (u'all', u'parsers'):
-      self._extraction_worker.SetParsersProfiler(None)
-      self._parsers_profiler.Write()
-      self._parsers_profiler = None
-
-    if self._profiling_type in (u'all', u'processing'):
-      self._extraction_worker.SetProcessingProfiler(None)
-      self._processing_profiler.Write()
-      self._processing_profiler = None
-
-    if self._profiling_type in (u'all', u'serializers'):
-      self._serializers_profiler.Write()
-      self._serializers_profiler = None
-
-  def SignalAbort(self):
-    """Signals the process to abort."""
-    self._abort = True
-    self._extraction_worker.SignalAbort()
-    self._parser_mediator.SignalAbort()
-
-
-class MultiProcessingQueue(plaso_queue.Queue):
-  """Class that defines the multi-processing queue."""
-
-  def __init__(self, maximum_number_of_queued_items=0, timeout=None):
-    """Initializes the multi-processing queue object.
-
-    Args:
-      maximum_number_of_queued_items (Optional[int]): maximum number of queued
-          items, where 0 represents no limit.
-      timeout (Optional[float]): number of seconds for the get to time out,
-          where None will block until a new item is put onto the queue.
-    """
-    super(MultiProcessingQueue, self).__init__()
-    self._timeout = timeout
-
-    # maxsize contains the maximum number of items allowed to be queued,
-    # where 0 represents unlimited.
-
-    # We need to check that we aren't asking for a bigger queue than the
-    # platform supports, which requires access to this protected member.
-    # pylint: disable=protected-access
-    queue_max_length = multiprocessing._multiprocessing.SemLock.SEM_VALUE_MAX
-    # pylint: enable=protected-access
-
-    if maximum_number_of_queued_items > queue_max_length:
-      logging.warning((
-          u'Requested maximum queue size: {0:d} is larger than the maximum '
-          u'size supported by the system. Defaulting to: {1:d}').format(
-              maximum_number_of_queued_items, queue_max_length))
-      maximum_number_of_queued_items = queue_max_length
-
-    # This queue appears not to be FIFO.
-    self._queue = multiprocessing.Queue(maxsize=maximum_number_of_queued_items)
-
-  def Open(self):
-    """Opens the queue."""
-    pass
-
-  # pylint: disable=arguments-differ
-  def Close(self, abort=False):
-    """Closes the queue.
-
-    This needs to be called from any process or thread putting items onto
-    the queue.
-
-    Args:
-      abort: optional boolean to indicate the close is issued on abort.
-    """
-    if abort:
-      # Prevent join_thread() from blocking.
-      self._queue.cancel_join_thread()
-
-    self._queue.close()
-    self._queue.join_thread()
-
-  def Empty(self):
-    """Empties the queue."""
-    try:
-      while True:
-        self._queue.get(False)
-    except Queue.Empty:
-      pass
-
-  def IsEmpty(self):
-    """Determines if the queue is empty."""
-    return self._queue.empty()
-
-  def PushItem(self, item, block=True):
-    """Pushes an item onto the queue.
-
-    Args:
-      block: boolean value to indicate put should block
-             if the queue is full.
-    """
-    try:
-      self._queue.put(item, block=block)
-
-    # Queue.Full can be raised if block is False.
-    # Since this should only be used in the abort code path
-    # the exception is ignored.
-    except Queue.Full:
-      pass
-
-  def PopItem(self):
-    """Pops an item off the queue.
-
-    Raises:
-      QueueClose: if the queue has already been closed.
-      QueueEmpty: if no item could be retrieved from the queue within the
-                  specified timeout.
-    """
-    try:
-      # If no timeout is specified the queue will block if empty otherwise
-      # a Queue.Empty exception is raised.
-      return self._queue.get(timeout=self._timeout)
-    except KeyboardInterrupt:
-      raise errors.QueueClose
-    # If close() is called on the multiprocessing.Queue while it is blocking
-    # on get() it will raise IOError.
-    except IOError:
-      raise errors.QueueClose
-    except Queue.Empty:
-      raise errors.QueueEmpty
